+++ src/ALPlanner/TrajectoryOptimizer/IKOrderCurve.cs
@@ -4,9 +4,9 @@ namespace ALPlanner.TrajectoryOptimizer.Curves;
 
 interface IKOrderCurve
 {
-    public IEnumerable<Vector3d> TrajectoryPoints(float fromWhen, float toWhen, float step);
+    public IEnumerable<Vector3d> TrajectoryPoints(double fromWhen, double toWhen, double step);
 
-    public Vector3d Value(float time);
+    public Vector3d Value(double time);
 
     public void Construction(IEnumerable<Vector3d> positionList, Vector3dTuple2 HeadTailVelocity, Vector3dTuple2 HeadTailAcceleration);
 }
